backend/requirements.txt
Django>=4.2
mongoengine>=0.27.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
